Update requirements and version in summarizer
Extended the list of requirements in summarizer/requirements.txt to improve the functionality and reliability of the summarizer. This included adding packages such as 'anyio', 'async-timeout', 'azure-common', 'black', and others. 'youtube_transcript_api' is changed to 'youtube-transcript-api' for consistency. Also, the version number of the application was incremented in src/manifest.json from 1.2